Backup
Made some of the game over reset work, now they respawn on the correct position, but it does not work fully yet as if they die in a different scene than they started in, the GameOverUI does not seem to be found on reset.
So I'm making a backup here in case further changes destroy the code.